Remove some CPP magic for runtime variables.
Previously, we were doing a #define for handling routing
this to the thread local data. This caused a problem when
a class had a slot named, say, %function, as that resulted
in a struct member 'Pfunction_' which then got defined to
something invalid.

This removed this issue entirely by removing those defines
and altering the code that is generated to match.

This bootstraps on OS X without an issue (and confirmed to fix
the compilation issue, which arose when building dylan-environment).