Fix an object corruption in <thread> instances in C-R-T
An incorrect assumption was made that DTHREAD can be extended.

To rectify this, I have introduced a per-thread GC-allocated THREAD
structure that contains thread-implementation-specific runtime details
about the thread.

This information should really go into the TEB, but that would
contaminate it with platform minutia.

Incidentally, this patch also keeps track of the thread name for further
use.