Remove some threading hacks for conditions in HARP.
This removes much of an old hack that turned *current-handlers*
into a storage slot in the TEB. This preserves keeping the value
in the TEB, but also stores it as a normal thread variable like
we do for the other backends.

We keep the TEB slot for now because the debugger needs it. In the
future, more of this hack can be removed.