Allocate exact amount of stack for vectors.
Previously, we had vectors of size 16 all of the time. Now, we use
variable length arrays to allocate memory for the vector and then
treat that memory as a vector.

This is better than using alloca() as it allows the memory to be freed
on exiting the scope rather than only at the end of the function.