Fix compiler crash: No applicable method for write.
We check to see if the binding is defined here so that we don't
end up with an error when using a symbol from another library
that is exported but undefined. The compiler has already
emitted warnings so we can just silently continue at this
point to avoid a compiler crash.

Closes #173.