[crt] Alter Boehm GC build integration.
This lets some platforms (Darwin) use the static linking method
while other platforms (Linux & FreeBSD) continue to require a
copy of Boehm to be installed in the system.

The static linking didn't work on x86_64 Linux because the static
library wasn't built with -fPIC, so it wouldn't link with our
shared library code.