Make <cpl-differ> a program-warning, rather than a serious-program-warning
Since 2012.1 is released, we have a release where the CPL difference
between DRM and C3 is a serious. According to discussions on hackers@
(especially https://lists.opendylan.org/pipermail/hackers/2011-December/006279.html)
we should remove the warning.

My plan is to ship next release with a warning, and afterwards
remove that entirely