[system] Correctly treat time_t as 64 bit.
On a 64 bit system, time_t is 64 bit, so we must
treat it as a long when reading / writing it.

This means we don't need the calloc hack as we're
now correctly writing the full value to memory.

This will also let this work with MPS on 64 bit
eventually as MPS doesn't clear memory pages
like the Boehm GC does.