Avoid executable stacks on Linux and FreeBSD.
Because we're generating assembly, we need for this section to
be present in the executable. We can also force it via the linker
flag.

Note that the generated .o files in the repo will need to be
regenerated with this section present at some point. (But the
linker flag keeps us safe for now.)