[rt] Use sigemptyset instead of sigfillset.
When using sigfillset, all signals are masked. Unfortunately,
this includes the signals that MPS requires, which leads to
deadlocks within MPS when it tries to stop the world.

We don't need to mask any signals here, so using sigemptyset
seems safe and is what the code did here before this code was
added.