llvm-as: New standalone tool with llvm-builder output
This commit adds a Dylan version of the LLVM assembler tool,
llvm-as. In addition to being useful for testing the Dylan llvm
library's bitcode output, it can also be used to generate (an
approximation of) Dylan builder code which can be adapted for use in
compiler and runtime support code generators.

* sources/registry/generic/llvm-as: New registry entry.

* sources/app/llvm-as/llvm-as.lid: New library interchange file.

* sources/app/llvm-as/library.dylan: New library/module definition source.

* sources/app/llvm-as/llvm-as.dylan: New main program for llvm-as.

* sources/app/llvm-as/llvm-write-builder.dylan: New source (builder
  output).

* sources/lib/llvm/llvm-library.dylan
  (module llvm): Export instruction slot getter functions needed to
   inspect instructions for builder call generation.