[crt,c-back-end] Switch to using "dylan_value" instead of "D" for Dylan values.
A Dylan value is a tagged value and may be an integer, byte character,
unicode character or a dylan_object.

This change is useful for working with lldb as the currently shipping
version of lldb crashes when adding a summary function for 'D'.

The type "D" is maintained for 2 reasons: just in case, for backwards
compatibility and in some situations it is used as the type for a
raw pointer. We will shift that over to "void *" in subsequent
commits.