[notebook] Stub version of Project Notebook.
This is where we'll record information about various projects
for Open Dylan. This would include future extensions or changes,
work that is ongoing, things that we'd like to clean up, research
into particularly troublesome issues, etc.

This isn't all stuff that belongs in the bug tracker as that's
being used for actual bugs.

It doesn't really belong in the wiki as that doesn't allow for
the description and write-ups to remain in sync with the code
as it changes / evolves.