[build] Move build scripts to their own directory.
This removes the build scripts from the 'lib' directory and places
them in a more appropriate 'share/opendylan/build-scripts' directory
(on Unix, on Windows, the path is just 'build-scripts') under the
installation target directory.

In the future, we may consider moving config.jam to a different
directory as it can contain architecture-dependent information.