[configure] Default to doing a 3-stage-bootstrap.
For anyone doing a build out of the box with the intent of
installing it, a 3-stage-bootstrap is required.

A common source of confusion for people is that they need to
type something more than 'make' for what is the required
course of action.

Those who know what they are doing and know that a shorter
build is possible to do correctly can use 'make 1-stage-bootstrap'
as they want.