[dfmc] Fix for passing struct by value in C backend.
This is a temporary fix for passing structs by value when using
C-FFI with the C backend. This allows for code to be generated
and compile.

It isn't ideal as it would be nice to support "c-name:" options
in C-struct-definer and pass that along and allow it to be used
here, but that's a future improvement.