[c-ffi,dfmc-back-end] Fix nested structs by value.
In maybe-claim-heap-element-references-internal, when processing
a raw aggregate type (used for values whose type is a C struct
or union and not a pointer to a struct/union), process the member
types first, if necessary, then the type itself.

This ensures that they're in the heap-referenced-objects list
in the correct order and then emitted in C in the correct order
in emit-typedefs in the C backend.  Without doing this, the C
output will be incorrect and compilation will fail.