[interface-reference] Emit better method docs.
If there's a single method on a sealed generic, use the type
signature from the method. This makes for better (and more
accurate docs) in many cases, like for slot getters and setters
as well as implicitly defined generics.

If a generic has multiple methods, emit the method specializations
after the documentation for the generic.