[dfmc] Propagate not-inline to generated code.
When a function is flagged as not-inline in Dylan, it is useful for
that information to be passed along to C and LLVM to prevent it from
being inlined during code generation.

This is particularly useful in tests where you want to prevent the
compiler from optimizing away the thing that you're testing.

* sources/dfmc/llvm-back-end/llvm-emit-code.dylan
  (llvm-function-attributes): New.
  (emit-code): Get function attributes and pass them along to the
   <llvm-function>.

* sources/dfmc/c-back-end/c-emit-object.dylan
  (emit-lambda-interface-using-function): Pass not-inline along to
   the generated C function.

* sources/lib/run-time/run-time.h
  (OPEN_DYLAN_NO_INLINE): New portability macro.