[dfmc-c-back-end] Fix generated C for infinity and NaN values.
Fixes #876.

* sources/dfmc/c-back-end/c-emit-object.dylan
  (print-raw-object on <float>): Examine the result of float-to-string
    and generate correct C for infinite and NaN values by looking for
    the string representations that float-to-string creates.

This examines the string representation because classify-float hasn't
landed yet, and even if it had, it wouldn't be available here yet
until we'd done a release due to bootstrapping issues.

Infinite and NaN values aren't commonly present as values that need
to have C generated, but it can happen. One primary cause of this
happening is the lack of floating point exception handling in the
C back-end and run-time, so constant folding within the compiler
can result in infinite and NaN values. This will be addressed in
future work.