[dfmc-llvm-back-end] Add %running-dylan-spy-function?.
%running-dylan-spy-function? is a variable which indicates
whether that the debugger is executing a function via the spy thread.
By default it is set to 0. The Dylan debugger will set it to something
else if it is executing code within the app.

The HARP run-time provides this and some code within the collector
requires it to exist when building the MPS collector.

* sources/dfmc/llvm-back-end/llvm-primitives-debug.dylan
  (%running-dylan-spy-function?): Define new runtime-variable.