dylan: Fix size parameter handling in range
Current implementation of range gives preference to size over to/below
parameters and modifies to accordingly if needed. This modification is
not correct as current as test range-5 in apple dylan test suite shows.

This change fixes size's behaviour according to current implementation,
making every test case in range-5 pass.

* sources/dylan/range.dylan
  (make): Fix size parameter handling