dfmc-llvm-linker: Split out library self-initialization
The debugger needs to be able to track the initialization of a library
independent of the calls it makes to initialize the libraries it
references. This commit splits the calls to the for-user init functions
of each compile-unit in the library into a separate function so that
the debugger can set tracking breakpoints at its entry and return.

* sources/dfmc/llvm-linker/llvm-gluefile.dylan
  (emit-gluefile): Generate the self-init function containing calls to
   the for-user init functions, and add a call to the self-init function
   at the end of the library init funciton.

* documentation/hacker-guide/source/runtime/startup.rst: Document the
  operation of the self-init function.