Improve fragment presentation on sequence literals.
Previously, we just used the default literal fragment presentation
which takes the literal value and prints it in a single shot with %=.

This breaks things that need to print differently based upon the actual
fragments, like keyword vs symbol syntax for symbols. It also won't support
nice line wrapping and other pretty printing behavior should that be desired.

* sources/dfmc/reader/presentation.dylan
  (present on sequence literal fragments): Implement.

* sources/dfmc/reader/tests/literal-test-suite.dylan
  (test list-literal-test): Verify presentation of parsed fragments.
  (test pair-literal-test): Likewise.
  (test vector-literal-test): Likewise.