dfmc-c-back-end: Fix MV #rest type checks
This is a follow up to  d39243638c0eae479847a45fa6109caaa692a265.

When an mv-check-type-rest happens as a result of a function which
has no specified values in the signature, then the code prior to
this commit was not emitting any transfer of values. This led to
an uninitialized memory read when attempting to use the return
value from this function.

(This happened in the objective-c bridge library with the various
as-raw-* methods.)

Fixes #1016.

* sources/dfmc/c-back-end/c-emit-computation.dylan
  (emit-computation on <multiple-value-check-type-rest>): Add missing
   transfer of values when there is no value specification in the
   signature.