Mark the stream closed if creating the accessor signals a condition
The initialize method for <file-stream> attempted to create the accessor before allocating the internal buffer. If the accesspr signalled a condition (specifically, if it was created with `if-exists` = #"signal" then the stream would be left in an open state but with no buffer. This then causes an unexpected condition when the stream is closed, possibly during termination.