dfmc-reader: change #name: syntax to #:name:
The experimental #name: syntax, having no reserved initial character
after the sharp character, means that any new # syntax added in the
future risks creating backward incompatibilities with code that
depends on it. In addition, the names may conflict with #b, #o, #x, et
al.  This change is in preparation for allowing use of underscores in
numeric literal syntax.

This also adds vertical bar, backquote, and single quote as valid
delimiters in #:name: syntax.