system: Remove unused class-info variable
This call was invalid because <wrapper-stream> isn't a registered class
when the system test suite is run by itself. It probably works when
combined into one executable with the io test suite since <wrapper-stream>
is registered by the io test suite, but that dependency shouldn't be a
requirement to run system-test-suite.