common-dylan: Fix performance in join method.
replace-subsequence! call in join method misses the end argument
as the replacement size should be the same as sequence size.

If this is not explicitly indicated the whole result suffix will be
discarded, with a huge performance penalty for large sequences,
as this operation is done in the inner loop.