dfmc-llvm-back-end: Fix stack overflows from BEF allocation
Code generation for <bind-exit> computation stack-allocates a bind
exit frame (BEF) but did not formerly free the frame at the end of the
scope. This could cause stack overflow for <bind-exit> within a loop.

This change uses the llvm.stacksave and llvm.stackrestore intrinsics
to put the stack pointer back where it was at the head of the
<bind-exit> computation.

* sources/dfmc/llvm-back-end/llvm-emit-computation.dylan
  (emit-computation on <bind-exit>): Retrieve a stack pointer value
  from before the BEF is allocated. In both the fall-through and
  exception handling cases, restore the saved stack value after the
  BEF is no longer of use.