dylan: Fix thread finalization reference cycle problems
Return values from a thread function are stored within the <thread>
object so that they can be returned when the thread is joined. Since
the <thread> object is registered for finalization, if the return
values reference it (directly or indirectly) then this will cause a
reference cycle involving a finalized object. This change reduces the
risk of this happening by clearing the results slot once the thread
has been joined and the values have been returned.

* sources/dylan/thread.dylan
  (join-thread): After joining a thread, clear out the
   function-results slot.