Registry cleanups: remove sources/examples/dylan-programming
Removing libraries that are in the source tree in two places, as part of work
to make it possible to generate the registry automatically in order to be able
to build with packages.

These libraries are also in sources/examples/documentation/airport, which is
the one that's included in the library packs, so let's keep that one.