Fix undefined binding to format-to-string in life
The module definition of life is missing the use of module
simple-formating to get access to the method format-to-string.

The application then crashes when you run it and click on the board
to set the starting state of the game.

Adding the use of module simple-format in the module definition of
life solves the issue.