common-dylan: fix bug in copy-bytes(<vector>, <string>)
`write(<sequence-stream>, <string>)` ends up calling `copy-bytes(<vector>,
<string>)` when the underlying sequence is a `<vector>`.  `copy-bytes` isn't
documented, but I infer from the implementations that it is misnamed and should
be called `copy-elements` or `copy-sequence-into!`, and my fix makes that
assumption. That is, that it isn't just intended to copy bytes.

```
Before:
test-bug-1360 failed
  #[1, 2, 3, 'A', 'B', 'C'] = stream-contents(stream) failed [#[1, 2, 3, 'A', 'B', 'C'] and #[1, 2, 3, 65, 66, 67] are not =.  element 3 is the first non-matching element]

After:
test-bug-1360 passed in 0.000216 seconds with 26016 bytes allocated.
  #[1, 2, 3, 'A', 'B', 'C'] = stream-contents(stream) passed
```

Fixes #1360