GitHub Actions: Do a bootstrap build.
This is only on `ubuntu-20.04` for now.

The `macos-latest` bootstrap is not working with the batteries-included
download due to a broken C compiler.

On `ubuntu-22.04`, some of the numerics tests fail as conditions
aren't getting signaled for floating point errors. I don't know
why.

We're using the 2020.1 release for now for no real reason.