llvm-runtime-generator: Mark run-time routine debug info as C99
The debug information generated by llvm-runtime-generator for the
primitives and entry-points it writes out was previously marked as
using the DW_LANG_Mips_Assembler language tag. Since lldb now insists
that the debug information correspond to a language it recognizes
before it will interpret the stack frame, this commit changes the
language tag to DW_LANG_C99.

* sources/app/llvm-runtime-generator/llvm-runtime-generator.dylan
  (generate-runtime): Mark the debug compile unit for the generated
   runtime as $DW-LANG-C99.