llvm: Improve method dispatch when writing bitcode
Previously, most bitcode records were written using the write-record
function, which iterates through its #rest arguments and calls
write-vbr on each. This means that the compiler cannot determine which
write-vbr method will be applicable, increasing the dispatch
overhead. This change uses a write-record* macro to make significanly
more of the arguments in write-vbr calls visible to the compiler's
typing mechanism.

* sources/lib/llvm/bitcode.dylan
  (write-record): Factor out the record header writing into
   write-record-head.
  (write-record-head): New method for writing out the three VBRs at
   the beginning of unabbreviated records.
  (write-record*): New macro for writing unabbreviated records.
  (%count-operands): New macro for counting write-record* operands.
  (%write-operands): New macro for writing write-record* operands by
   expanding into calls to write-vbr.

* sources/lib/llvm/llvm-bitcode.dylan: Replace calls to write-record
  with invocations of write-record* wherever possible.
  (write-type-record on <llvm-primitive-type>,
   write-type-record on <llvm-integer-type>,