mps: Use gc_teb's allocation counter with HARP.
When building with LLVM and using MPS, still need to use the
LLVM code path for the allocation counter as the gc_teb doesn't
exist within the LLVM backend.

This isn't sufficient to get allocation working as other things
still use the gc_teb, but they can be fixed in subsequent commits.