documentation: Fix :macrocall: markup
The old way was to use `.. code-block:: dylan` followed by an example using `*`
where it wanted italics. But inside code-block `*` isn't recognized, so this
came out looking like lots of global variables.

The new way used `.. parsed-literal:: ` (note the trailing space) followed by
an example using backquotes to create italics. This has the additional
advantage that when editing the source in emacs the italicized parts are
colorized, but only when that trailing space is present in the parsed-literal
directive! Go figure. (There's some risk that Future Me will compulsively do
M-x delete-trailing-whitespace and break that.)