dfmc-reader: more multi-line string fixes
This is a reworking of the multi-line string implementation due to previous
misunderstandings of the spec.
https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/proposals/csharp-11.0/raw-string-literal
clarifies many things. There are two primary changes here:

(1) The content of a multi-line string begins after the first Newline and ends before the
last Newline. A consequence of this is that the empty string is written as:
    """

    """

(2) The start/end delimiters may be any length (but obviously must match each
other). This makes it possible to easily include snippets of code from any language in a
string, for example.  To manage this I introduce a new <conditionally-accepting-state>
that allows the lexer to be side-effected, thus making it possible to track the number of
double quotes in the start/end delimiters and decide when to accept the string token.