dfmc-reader: get-token cleanup and refactoring
This is a follow on to the previous commit, but in this case changing local variable
names in get-token; due to its size they should be exceptionally clear. The majority of
the changes are towards using "line-number" instead of just "line" and "foo-pos" instead
of just "foo".

Since we're now passing the lexer into get-token-1 again (it's a tragic story, don't ask)
there's no longer a need to pass all the args that are derived from the lexer's slots, so
don't.

There shouldn't be any logic changes in this commit, just simple substitutions.